centralize global app state
