+++ src/components/NpsSurvey.js
@@ -6,9 +6,9 @@ import { buttonSecondary, Clickable } from 'src/components/common'
 import { icon } from 'src/components/icons'
 import { TextArea } from 'src/components/input'
 import { Ajax } from 'src/libs/ajax'
-import { authStore } from 'src/libs/auth'
 import colors from 'src/libs/colors'
 import { getAppName } from 'src/libs/logos'
+import { authStore } from 'src/libs/state'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
