+++ src/components/TrialBanner.js
@@ -1,13 +1,13 @@
 import _ from 'lodash/fp'
 import { div, h, a, span } from 'react-hyperscript-helpers'
-import { authStore, refreshTerraProfile } from 'src/libs/auth'
+import { refreshTerraProfile } from 'src/libs/auth'
 import colors from 'src/libs/colors'
 import { buttonPrimary, Clickable } from 'src/components/common'
 import { icon } from 'src/components/icons'
-import { freeCreditsActive } from 'src/components/FreeCreditsModal'
 import { ajaxCaller } from 'src/libs/ajax'
 import { reportError } from 'src/libs/error'
 import { getAppName } from 'src/libs/logos'
+import { authStore, freeCreditsActive } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
 import { Component } from 'src/libs/wrapped-components'
 import Modal from 'src/components/Modal'
