+++ src/pages/billing/NewAccountModal.js
@@ -5,8 +5,8 @@ import { buttonPrimary, buttonSecondary, Checkbox, Clickable, Select } from 'src
 import { icon } from 'src/components/icons'
 import { ValidatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
-import { authStore } from 'src/libs/auth'
 import colors from 'src/libs/colors'
+import { authStore } from 'src/libs/state'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
 import validate from 'validate.js'
